Chapter2/pythonClientSample/requirements.txt
elasticsearch
elasticsearch-async
load_dotenv
tqdm
urllib3
